% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvSingle.R
\name{deconvSingle}
\alias{deconvSingle}
\title{Deconvolve the true gene expression distribution  of a single gene}
\usage{
deconvSingle(y, scaling.consts = NULL, Z = NULL, Z0 = NULL,
  plot.density = T, do.LRT.test = T, family = c("Poisson",
  "Negative Binomial"), NB.size = 100, verbose = T, control = list())
}
\arguments{
\item{y}{a vector of observed counts across cells for a single gene}

\item{scaling.consts}{a vector of cell specific scaling constants, either the cell efficiency or the library size}

\item{Z}{covariates for active intensity. Default is NULL.}

\item{Z0}{covariates for active fraction. Used only when zeroInflate is True. Default is NULL.}

\item{plot.density}{whether plot the density curve of the deconvolved the distribution or not. The zero inflation part has been smoothed into the density curve for visualization. Default is True.}

\item{do.LRT.test}{whether do LRT test on the coefficients and active fraction or not. Default is True}

\item{family}{family of the noise distribution, support either "Poisson" or "Negative Binomial" with known tuning parameter}

\item{NB.size}{over-dispersion parameter when the family is Negative Binomial: mu = mu + mu^2/size}

\item{verbose}{verbose the estimation and testing procedures or not. Default is True.}

\item{control}{settings see {\code{\link{DESCEND.control}}}}
}
\value{
a DESCEND object. See also \code{\link{DESCEND}}
}
\description{
The deconvolution is computed by using the function {\code{\link{deconvG}}}. This function can automatically discretize the underlying distribution and find the proper tuning parameter \code{c0} of the penalty term. Besides, it computes the estimates and standard deviations of five distribution based statistics (active fraction, active intensity, mean, CV and gini coefficient), as well as the estimated coefficients of the covariates on active intensity (Z) and active fraction (Z0), and store them in a DESCEND object.
}
\examples{

X <- rpois(1000, 0.2 * 3)
Z <- rnorm(1000)
result <- deconvSingle(X, Z = Z, Z0 = Z, scaling.consts = rep(0.2, 1000), do.LRT.test = TRUE)
result@estimates
}
