% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descend.R
\name{descend}
\alias{descend}
\title{Apply descend to each gene of the count matrix}
\usage{
descend(count.matrix, scaling.consts = NULL, Z = NULL, Z0 = NULL,
  n.cores = 1, cl = NULL, do.LRT.test = F, family = c("Poisson",
  "Negative Binomial"), NB.size = 100, verbose = T, control = list())
}
\arguments{
\item{count.matrix}{the observed UMI count matrix. Each row is a gene and each column is a cell}

\item{scaling.consts}{a vector of cell specific scaling constants, either the cell efficiency or the library size}

\item{Z}{covariates for active intensity. Default is NULL.}

\item{Z0}{covariates for active fraction. Used only when zeroInflate is True. Default is NULL.}

\item{n.cores}{the number of cores used for parallel computing. Default is 1. Used only when parallel computing is done in a single machine. For using multi-machine cores, need to assign \code{cl} explicitly}

\item{cl}{an object of class "cluster". See more details in \code{\link[parallel]{makeCluster}}}

\item{do.LRT.test}{whether do LRT test on the coefficients and active fraction or not. Default is True}

\item{family}{family of the noise distribution, support either "Poisson" or "Negative Binomial" with known tuning parameter}

\item{NB.size}{over-dispersion parameter when the family is Negative Binomial: mu = mu + mu^2/size}

\item{verbose}{verbose the estimation and testing procedures or not. Default is True.}

\item{control}{settings see see {\code{\link{deconv.control}}}}
}
\value{
a list of DESCEND objects
}
\description{
Apply descend to each gene of the count matrix
}
