% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{deTest}
\alias{deTest}
\title{Perform differential testing of the distribution measurements between two cell populations}
\usage{
deTest(descend.multipop.output, de.pair.names, count.matrix, labels,
  N.genes.null = 10000, alternative = c("two.sided", "less", "greater"),
  n.cores = 1, cl = NULL, verbose = T)
}
\arguments{
\item{descend.multipop.output}{the returned value of \code{\link{descendMultiPop}}}

\item{de.pair.names}{a vector of length 2 showing the two cell population names for differential testing. The names should match the values in \code{labels}}

\item{count.matrix}{the observed UMI count matrix. Each row is a gene and each column is a cell. The column sums are used as the input for \code{scaling.consts} when both \code{ercc.matrix} and \code{scaling.consts} are NULL.}

\item{labels}{a vector of factors or characters, indicating the cell popluation label of each cell. The length of \code{labels} should be the same as the number of columns of \code{count.matrix}}

\item{N.genes.null}{number of permuted genes. The larger the value is, the longer it takes and the smaller the minimum p-value can be (the minimum p-value can be as small as 0.5/N.genes.null)}

\item{alternative}{the alternative hypotheses for differential testing}

\item{n.cores}{the number of cores used for parallel computing. Default is 1. Used only when parallel computing is done in a single machine. For using multi-machine cores, need to assign \code{cl} explicitly}

\item{cl}{an object of class "cluster". See more details in \code{\link[parallel]{makeCluster}}}

\item{verbose}{verbose the estimation and testing procedures or not. Default is True.}
}
\value{
a list with elements
\item{p.values}{a matrix of p-values calculated from permutation tests. Each row is a gene and each column is a distribution measurement of coefficient (if \code{Z} or \code{Z0} is presented)}
\item{scores}{the z-scores matrix in differential testing. Each row is a gene and each column is a distribution measurement of coefficient (if \code{Z} or \code{Z0} is presented)}
\item{perm.scores}{the permuted z-scores matrix used as the null distribution of the z-scores in each column. The number of rows equals to \code{N.genes.null}}
}
\description{
Permutation test is used to get the p-values in differential testing. To keep the possible correlation structure between genes. We run \code{ceiling(N.genes.null/nrow(count.matrix))} rounds of permutation. For each round, we permute cell label once and deconvolve the true expression distribution for all the selected genes with the permuted cell labels. One may need more rounds of permutation to get smaller p-values. See \code{\link{deTest.more}}.
}
