% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descend.R
\name{findHVG}
\alias{findHVG}
\title{Finding highly variable genes (HVG) based on Gini coefficients or CV}
\usage{
findHVG(descend.list, criteria = c("Gini", "CV"), quantile = 0.7,
  threshold = 3, plot.result = T, spline.df = 5)
}
\arguments{
\item{descend.list}{a list of descend objects computed from {\code{\link{runDescend}}}}

\item{criteria}{the cirteria of HVG finding, either "Gini" or "CV"}

\item{quantile}{quantile of the quantile regression. Defined as the \code{tau} parameter in the function \code{\link[quantreg]{rqss}}. A larger quantile yields a more stringent selection of HVG. Default is 0.7.}

\item{threshold}{threshold of the normalized difference for HVG selection. Default is 3. A larger value results in a more stringent selection}

\item{plot.result}{whether plot the selection results or not. Default is TRUE.}

\item{spline.df}{the degree of freedom of the spline functions used to fit the quantile regression curve}
}
\value{
A list of elements:
\item{score.mat}{A score matrix of the genes. Each row is a gene}
\item{HVG.genes}{A vector of the selected HVG gene names}
}
\description{
The function finds highly variable genes based on the Gini or CV estimates from DESCEND. A quantile natural cubic spline regression of the estimated selected criteria on the log of the estimated mean for the genes is performed as the general trend. HVGs are selected as the genes whose normalized difference between its estimates and the trend is larger than a threshold. For examples, see \code{\link{runDescend}}
}
