% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descend.R
\name{getEstimates}
\alias{getEstimates}
\title{Grab the value and standard deviation of the estimated parameters calculated from a list of descend objects}
\usage{
getEstimates(descend.list)
}
\arguments{
\item{descend.list}{a list of descend objects}
}
\description{
Grab the value and standard deviation of the estimated parameters calculated from a list of descend objects
}
