% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descend.R
\name{getPval}
\alias{getPval}
\title{Grab the likelihood ratio test p-values of the coefficients and active fraction is the tests are performed from a list of descend objects}
\usage{
getPval(descend.list)
}
\arguments{
\item{descend.list}{a list of descend objects}
}
\description{
Grab the likelihood ratio test p-values of the coefficients and active fraction is the tests are performed from a list of descend objects
}
