% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvSingle.R
\name{DESCEND}
\alias{DESCEND}
\alias{DESCEND-class}
\title{An S4 class object containing the DESCEND result for a single gene}
\description{
The DESCEND class is a container to store the DESCEND result for a single gene.
}
\section{Slots}{

\describe{
\item{\code{distribution}}{The distribution of the deconvolved distribution with relative statistics}
\item{\code{estimates}}{The matrix of distribution measurements and coefficients estimated values, bias, standard deviation and mean square error (named as \code{DESCEND.sd})}
\item{\code{pval}}{The p values of the likelihood ratio tests if computed}
\item{\code{density.points}}{Smoothed version of the distribution for easier plotting of the distribution}
}
}

\author{
Jingshu Wang
}
