% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descend.R
\name{getEstimates}
\alias{getEstimates}
\title{Grab the value and standard deviation of the estimated DESCEND elements calculated from a list of descend objects. For examples, see  \code{\link{runDescend}}}
\usage{
getEstimates(descend.list)
}
\arguments{
\item{descend.list}{a list of descend objects computed from {\code{\link{runDescend}}}}
}
\value{
A list of matrices where each element is for a distribution measurement or a coefficient if covariates are presented. For each distribution measurement or coefficient, the matrix contains two columns. Each row is for a gene. The first column is the estimated value and second value is the estimated standard deviation.
}
\description{
Grab the value and standard deviation of the estimated DESCEND elements calculated from a list of descend objects. For examples, see  \code{\link{runDescend}}
}
